requirements.txt
actionlib==1.14.0
angles==1.9.13
asgiref==3.8.1
backports.zoneinfo==0.2.1
base-local-planner==1.17.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.8.30
charset-normalizer==3.4.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
diagnostic-updater==1.11.0
Django==4.2.16
djangorestframework==3.15.2
drone-wrapper==1.4.2
dynamic-reconfigure==1.7.3
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
idna==3.10
image-geometry==1.16.2
interactive-markers==1.12.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
laser-geometry==1.6.7
mavros==1.19.0
message-filters==1.16.0
moveit-commander==1.1.14
moveit-core==1.1.14
moveit-ros-planning-interface==1.1.14
moveit-ros-visualization==1.1.14
pr2-power-board==1.1.10
py-trees==0.7.6
python-qt-binding==0.4.4
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
resource-retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-controller-manager==0.20.0
rqt-dep==0.4.12
rqt-drone-teleop==1.4.2
rqt-ez-publisher==0.6.1
rqt-graph==0.4.14
rqt-ground-robot-teleop==1.4.2
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-joint-trajectory-controller==0.22.0
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-multiplot==0.0.12
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-pr2-dashboard==0.4.1
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-py-trees==0.4.1
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.25
sensor-msgs==1.13.1
smach==2.5.2
smclib==1.8.6
sqlparse==0.5.2
srdfdom==0.6.4
tf==1.13.2
tf-conversions==1.13.2
tf2-client==1.0.0
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tf2-sensor-msgs==0.7.7
tf2-server==1.1.3
topic-tools==1.16.0
typing-extensions==4.12.2
unique-id==1.0.6
urdfdom-py==0.4.6
urllib3==2.2.3
xacro==1.14.18
